+++ core/src/main/java/com/cmeza/spring/jdbc/repository/mappers/classes/DefaultJdbcRowMapper.java
@@ -1,4 +1,4 @@
-package com.cmeza.spring.jdbc.repository.mappers;
+package com.cmeza.spring.jdbc.repository.mappers.classes;
 
 public class DefaultJdbcRowMapper<T> extends JdbcRowMapper<T> {
     protected DefaultJdbcRowMapper(Class<T> mappedClass) {