+++ core/src/main/java/com/cmeza/spring/jdbc/repository/mappers/projections/DefaultJdbcProjectionRowMapper.java
@@ -1,4 +1,4 @@
-package com.cmeza.spring.jdbc.repository.projections;
+package com.cmeza.spring.jdbc.repository.mappers.projections;
 
 public class DefaultJdbcProjectionRowMapper<T> extends JdbcProjectionRowMapper<T> {
     protected DefaultJdbcProjectionRowMapper(Class<T> mappedClass) {