+++ core/src/main/java/com/cmeza/spring/jdbc/repository/mappers/records/DefaultJdbcRecordRowMapper.java
@@ -1,0 +1,7 @@
+package com.cmeza.spring.jdbc.repository.mappers.records;
+
+public class DefaultJdbcRecordRowMapper<T extends Record> extends JdbcRecordRowMapper<T> {
+    protected DefaultJdbcRecordRowMapper(Class<T> mappedClass) {
+        super(mappedClass);
+    }
+}