+++ core/src/main/java/com/cmeza/spring/jdbc/repository/projections/ProjectionAttributes.java
@@ -1,9 +1,0 @@
-package com.cmeza.spring.jdbc.repository.projections;
-
-import java.util.Map;
-
-public interface ProjectionAttributes {
-    ProjectionAttributes addAttribute(String name, Object value);
-
-    Map<String, Object> getAttributes();
-}