+++ core/src/test/java/com/cmeza/spring/jdbc/repository/records/TitleRecord.java
@@ -1,0 +1,10 @@
+package com.cmeza.spring.jdbc.repository.records;
+
+import java.util.Date;
+
+public record TitleRecord(
+        Integer employeeId,
+        String title,
+        Date fromDate,
+        Date toDate) {
+}