+++ core/src/test/java/com/cmeza/spring/jdbc/repository/repositories/informix/query/InformixQueryRepository.java
@@ -1,22 +1,24 @@
 package com.cmeza.spring.jdbc.repository.repositories.informix.query;
 
-import com.cmeza.spring.jdbc.repository.support.annotations.JdbcRepository;
-import com.cmeza.spring.jdbc.repository.support.annotations.methods.operations.JdbcQuery;
-import com.cmeza.spring.jdbc.repository.support.annotations.methods.supports.JdbcJoinTable;
-import com.cmeza.spring.jdbc.repository.support.annotations.methods.supports.JdbcMapping;
-import com.cmeza.spring.jdbc.repository.support.annotations.parameters.JdbcParam;
-import com.cmeza.spring.jdbc.repository.constants.TestConstants;
 import com.cmeza.spring.jdbc.repository.configurations.InformixInitializer;
+import com.cmeza.spring.jdbc.repository.constants.TestConstants;
 import com.cmeza.spring.jdbc.repository.mappers.EmployeeAndSalaryRowMapper;
 import com.cmeza.spring.jdbc.repository.mappers.EmployeeRowMapper;
 import com.cmeza.spring.jdbc.repository.mappers.projections.EmployeeAndTitleProjectionRowMapper;
+import com.cmeza.spring.jdbc.repository.mappers.records.EmployeeAndSalaryRecordRowMapper;
 import com.cmeza.spring.jdbc.repository.models.Department;
 import com.cmeza.spring.jdbc.repository.models.DepartmentEmployee;
 import com.cmeza.spring.jdbc.repository.models.Employee;
 import com.cmeza.spring.jdbc.repository.projections.EmployeeAndSalaryProjection;
 import com.cmeza.spring.jdbc.repository.projections.EmployeeAndTitleProjection;
 import com.cmeza.spring.jdbc.repository.projections.EmployeeProjection;
+import com.cmeza.spring.jdbc.repository.records.EmployeeRecord;
 import com.cmeza.spring.jdbc.repository.repositories.contracts.QueryContract;
+import com.cmeza.spring.jdbc.repository.support.annotations.JdbcRepository;
+import com.cmeza.spring.jdbc.repository.support.annotations.methods.operations.JdbcQuery;
+import com.cmeza.spring.jdbc.repository.support.annotations.methods.supports.JdbcJoinTable;
+import com.cmeza.spring.jdbc.repository.support.annotations.methods.supports.JdbcMapping;
+import com.cmeza.spring.jdbc.repository.support.annotations.parameters.JdbcParam;
 
 import java.sql.Types;
 import java.util.*;
@@ -114,4 +116,15 @@ public interface InformixQueryRepository extends QueryContract {
     @Override
     @JdbcQuery(table = "department", where = "id = :id")
     Optional<Department> getDepartmentOptionalWithCondition(String id);
+
+    @Override
+    @JdbcMapping(from = "conditionOne", to = "name", type = Types.VARCHAR)
+    @JdbcMapping(from = "conditionTwo", to = "gen", type = Types.VARCHAR)
+    @JdbcQuery(table = "employee", where = "first_name = :name and gender = :gen")
+    EmployeeRecord getOneEmployeeRecordWithConditionMapping(String conditionOne, String conditionTwo);
+
+    @Override
+    @JdbcJoinTable(table = "salary", alias = "s", on = "s.employee_id = e.id", join = JdbcJoinTable.Join.INNER)
+    @JdbcQuery(table = "employee", alias = "e", where = "e.id = :id", mapper = EmployeeAndSalaryRecordRowMapper.class)
+    EmployeeRecord getOneEmployeeRecordWithConditionAndComplexRowMapper(Integer id);
 }