+++ core/src/test/java/com/cmeza/spring/jdbc/repository/tests/contracts/QueryTestContract.java
@@ -41,4 +41,8 @@ public interface QueryTestContract {
     void testGetAllEmployeeWithConditionListAndMapping();
 
     void testGetDepartmentOptionalWithCondition();
+
+    void testGetOneEmployeeRecordWithConditionMapping();
+
+    void testGetOneEmployeeRecordWithConditionAndComplexRowMapper();
 }